requirements.txt
absl-py==0.10.0
anyio==4.0.0
appnope==0.1.0
argon2-cffi==20.1.0
arrow==1.3.0
astroid==2.4.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autopep8==1.5.4
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==3.1.5
boto==2.49.0
boto3==1.14.43
botocore==1.17.43
CacheControl==0.12.6
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
charset-normalizer==3.3.0
click==7.1.2
colorama==0.4.6
comm==0.1.4
cycler==0.10.0
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
download==0.3.5
entrypoints==0.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.0.12
firebase-admin==4.3.0
firestore==0.0.8
fqdn==1.5.1
ftfy==5.8
future==0.18.2
gast==0.3.3
google-api-core==1.22.0
google-api-python-client==1.10.1
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.30.0
google-cloud-vision==1.0.0
google-crc32c==0.1.0
google-pasta==0.2.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.59.0
grpcio-status==1.48.2
httplib2==0.18.1
idna==2.10
ipykernel==6.25.2
ipyplot==1.1.1
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isoduration==20.11.0
isort==4.3.21
jedi==0.17.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==3.0.2
msgpack==1.0.7
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
plac==1.1.3
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.6.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.1
pylint==2.5.3
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytokenizations==0.6.0
pytz==2023.3.post1
pyvisionproductsearch==0.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2020.7.14
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.4
rsa==4.9
s3transfer==0.3.7
sacremoses==0.0.43
Send2Trash==1.8.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
termcolor==1.1.0
terminado==0.17.1
testpath==0.4.4
tinycss2==1.2.1
toml==0.10.2
torchcontrib==0.0.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
types-python-dateutil==2.8.19.14
tzdata==2023.3
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.25.11
wasabi==0.7.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.15.0
